cal by rating

+++ new_cf/dataset.py
@@ -28,7 +28,7 @@ class Dataset:
         self.transaction = list(open(data_dir + "review_info.txt"))
         cols = self.transaction[0].split(', ')[:3]
         self.transaction = [i.strip().split(', ')[:3] for i in self.transaction[1:]]
-        self.transaction = pd.DataFrame(self.transaction, columns=cols)
+        self.transaction = pd.DataFrame(self.transaction, columns=cols).astype('int')
         self.transaction['train'] = False
         for i in range(self.no_user):
             self.transaction[self.transaction.u_id == i & self.transaction.p_id.isin(self.train[i])].train = True
